@@ -2,14 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:currency_decimal_formatter/currency_decimal_formatter.dart';
 
-void main() => runApp(MyApp());
+void main() => runApp(const MyApp());
 
 class MyApp extends StatelessWidget {
+  const MyApp({super.key});
+
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: ExamplePage(title: 'ExamplePage'),
+      home: ExamplePage(title: 'Currency Decimal Formatter Example'),
     );
   }
 }
@@ -17,18 +19,21 @@ class MyApp extends StatelessWidget {
 class ExamplePage extends StatefulWidget {
   final String title;
 
-  const ExamplePage({Key key, this.title}) : super(key: key);
+  const ExamplePage({super.key, required this.title});
 
   @override
-  _ExamplePageState createState() => _ExamplePageState();
-}
-
-class _ExampleMask {
-  final TextEditingController textController = TextEditingController();
+  State<ExamplePage> createState() => _ExamplePageState();
 }
 
 class _ExamplePageState extends State<ExamplePage> {
-  String _currency;
+  String _currency = 'USD';
+  final TextEditingController _textController = TextEditingController();
+
+  @override
+  void dispose() {
+    _textController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -38,26 +43,59 @@ class _ExamplePageState extends State<ExamplePage> {
       locale: locale,
       decimalDigits: decimalDigits,
     );
+    
     return Scaffold(
-        body: Center(
-            child: Row(children: [
-      DropdownButton<String>(
-        items: <String>['USD', 'EUR', 'JPY', 'BRL'].map((String value) {
-          return new DropdownMenuItem<String>(
-            value: value,
-            child: new Text(value),
-          );
-        }).toList(),
-        onChanged: (String value) {
-          _currency = value;
-        },
+      appBar: AppBar(
+        title: Text(widget.title),
+        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
       ),
-      TextField(
-        controller: textController,
-        inputFormatters: [
-          textFormatter,
-        ],
-      )
-    ])));
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                const Text('Currency: '),
+                const SizedBox(width: 8),
+                DropdownButton<String>(
+                  value: _currency,
+                  items: <String>['USD', 'EUR', 'JPY', 'BRL'].map((String value) {
+                    return DropdownMenuItem<String>(
+                      value: value,
+                      child: Text(value),
+                    );
+                  }).toList(),
+                  onChanged: (String? value) {
+                    if (value != null) {
+                      setState(() {
+                        _currency = value;
+                      });
+                    }
+                  },
+                ),
+              ],
+            ),
+            const SizedBox(height: 20),
+            TextField(
+              controller: _textController,
+              inputFormatters: [textFormatter],
+              decoration: const InputDecoration(
+                border: OutlineInputBorder(),
+                labelText: 'Enter currency amount',
+                hintText: '0.00',
+              ),
+              keyboardType: TextInputType.number,
+            ),
+            const SizedBox(height: 20),
+            Text(
+              'Selected Currency: $_currency',
+              style: Theme.of(context).textTheme.titleMedium,
+            ),
+          ],
+        ),
+      ),
+    );
   }
 }
